requirements.txt
pandas
requests
google-cloud-storage
openai
google-generativeai
python-dotenv
sendgrid
